# 42_DigiConnect/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/geography/tree": {
      "get": {
        "operationId": "GeographyController_getTree",
        "parameters": [
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get visual hierarchical jurisdiction tree",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/geography/nodes/{id}": {
      "get": {
        "operationId": "GeographyController_getNode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get jurisdiction node details by ID",
        "tags": [
          "geography"
        ]
      },
      "put": {
        "operationId": "GeographyController_updateNode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN or CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update jurisdiction node",
        "tags": [
          "geography"
        ]
      },
      "delete": {
        "operationId": "GeographyController_deleteNode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN or CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete jurisdiction node (fails if children exist)",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/geography/nodes/{id}/children": {
      "get": {
        "operationId": "GeographyController_getChildren",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get direct child jurisdiction nodes",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/geography/nodes/{id}/ancestors": {
      "get": {
        "operationId": "GeographyController_getAncestors",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get ancestor path up to the State root",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/geography/scope-check": {
      "get": {
        "operationId": "GeographyController_checkScope",
        "parameters": [
          {
            "name": "leafNodeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignedNodeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Check if a leaf node falls within an assigned officer node scope",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/geography/nodes": {
      "post": {
        "operationId": "GeographyController_createNode",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN or CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create new dynamic jurisdiction node",
        "tags": [
          "geography"
        ]
      }
    },
...
    "/api/v1/state-admin/departments": {
      "get": {
        "operationId": "StateAdminController_listDepartments",
        "parameters": [
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-state-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List departments in state",
        "tags": [
          "state-admin"
        ]
      },
      "post": {
        "operationId": "StateAdminController_createDepartment",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN or CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "State Admin creates department & appoints Department Head",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/departments/{id}": {
      "get": {
        "operationId": "StateAdminController_getDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get department details by ID",
        "tags": [
          "state-admin"
        ]
      },
      "put": {
        "operationId": "StateAdminController_updateDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update department details",
        "tags": [
          "state-admin"
        ]
      },
      "delete": {
        "operationId": "StateAdminController_deleteDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete department (fails if services exist)",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/grievance-cells": {
      "post": {
        "operationId": "StateAdminController_configureGrievanceCell",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureGrievanceCellDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Configure department grievance cell & workflow",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/grievance-cells/{deptId}": {
      "get": {
        "operationId": "StateAdminController_getGrievanceCell",
        "parameters": [
          {
            "name": "deptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get grievance cell configuration for department",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/analytics/revenue": {
      "get": {
        "operationId": "StateAdminController_getDepartmentRevenue",
        "parameters": [
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-state-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Department-wise revenue analysis",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/analytics/kpis": {
      "get": {
        "operationId": "StateAdminController_getStateAnalytics",
        "parameters": [
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-state-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "State Admin KPI overview",
        "tags": [
          "state-admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "UCSDP Backend API",
    "description": "Unified Citizen Service and Delivery Platform API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateNodeDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateNodeDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDepartmentDto": {
        "type": "object",
        "properties": {}
      },
      "ConfigureGrievanceCellDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}